Update json with login

+++ src/main/java/ua/nure/illiashenko/mutuallearning/service/TestService.java
@@ -1,6 +1,7 @@
 package ua.nure.illiashenko.mutuallearning.service;
 
 import static ua.nure.illiashenko.mutuallearning.constants.ArticleUserRole.ARTICLE_CREATOR;
+import static ua.nure.illiashenko.mutuallearning.constants.SystemUserRole.PREMIUM_USER;
 
 import java.sql.Timestamp;
 import java.util.ArrayList;
@@ -29,12 +30,15 @@ import ua.nure.illiashenko.mutuallearning.dto.test.question.answer.SaveAnswerReq
 import ua.nure.illiashenko.mutuallearning.entity.Answer;
 import ua.nure.illiashenko.mutuallearning.entity.Question;
 import ua.nure.illiashenko.mutuallearning.entity.Test;
+import ua.nure.illiashenko.mutuallearning.entity.User;
+import ua.nure.illiashenko.mutuallearning.entity.UserArticle;
 import ua.nure.illiashenko.mutuallearning.entity.UserTest;
 import ua.nure.illiashenko.mutuallearning.exception.ServiceApiException;
 import ua.nure.illiashenko.mutuallearning.repository.AnswerRepository;
 import ua.nure.illiashenko.mutuallearning.repository.QuestionRepository;
 import ua.nure.illiashenko.mutuallearning.repository.TestRepository;
 import ua.nure.illiashenko.mutuallearning.repository.UserArticleRepository;
+import ua.nure.illiashenko.mutuallearning.repository.UserRepository;
 import ua.nure.illiashenko.mutuallearning.repository.UserTestRepository;
 
 @Slf4j
@@ -51,8 +55,11 @@ public class TestService {
     private QuestionRepository questionRepository;
     @Autowired
     private UserArticleRepository userArticleRepository;
+    @Autowired
+    private UserRepository userRepository;
 
-    public void createTest(int articleId, SaveTestRequest saveTestRequest) {
+    public void createTest(String login, int articleId, SaveTestRequest saveTestRequest) {
+        checkAccessByArticleId(login, articleId);
         final Test test = new Test();
         test.setArticleId(articleId);
         test.setTitle(saveTestRequest.getTitle());
@@ -73,9 +80,10 @@ public class TestService {
         }
     }
 
-    public void editTest(int id, SaveTestRequest saveTestRequest) {
-        final Test test = new Test();
-        test.setId(id);
+    public void editTest(String login, int id, SaveTestRequest saveTestRequest) {
+        checkAccessByTestId(login, id);
+        final Test test = testRepository.findById(id)
+            .orElseThrow(() -> new ServiceApiException(List.of("testNotFound"), HttpStatus.NOT_FOUND));
         test.setTitle(saveTestRequest.getTitle());
         testRepository.save(test);
         deleteOldQuestionsWithAnswers(id, saveTestRequest);
@@ -94,6 +102,7 @@ public class TestService {
                     Collectors.toList()));
                 answerRepository.deleteAllById(oldAnswersId);
             }
+            question.setTestId(test.getId());
             question.setText(saveQuestionRequest.getText());
             question.setType(saveQuestionRequest.getType());
             final Question savedQuestion = questionRepository.save(question);
@@ -125,7 +134,8 @@ public class TestService {
         }
     }
 
-    public TestToUpdateResponse getTestToUpdate(int id) {
+    public TestToUpdateResponse getTestToUpdate(String login, int id) {
+        checkAccessByTestId(login, id);
         final Test test = testRepository.findById(id)
             .orElseThrow(() -> new ServiceApiException(HttpStatus.NOT_FOUND));
         final QuestionToUpdateResponse[] questionResponses = questionRepository.findByTestId(id)
@@ -158,9 +168,12 @@ public class TestService {
             .orElseThrow(() -> new ServiceApiException(HttpStatus.NOT_FOUND));
         final Integer maxMark = getMaxMark(id);
         final PreviousAttemptsResponse[] ownPreviousAttempts = getOwnPreviousAttempts(login, id);
-        final String role = userArticleRepository.findByUserLoginAndArticleId(login, test.getArticleId())
-            .getRole();
-
+        final Optional<UserArticle> optionalUserArticle = userArticleRepository.findByUserLoginAndArticleId(login,
+            test.getArticleId());
+        String role = "";
+        if (optionalUserArticle.isPresent()) {
+            role = optionalUserArticle.get().getRole();
+        }
         int userCount = 0;
         int sumMarks = 0;
         final List<UserTest> userTests = userTestRepository.findByTestIdOrderByUserLoginAscDateTimeDesc(id);
@@ -173,7 +186,7 @@ public class TestService {
                 currentUserTest = userTest;
                 userCount++;
             } else if (currentUserTest != userTest) {
-                if (isPremiumCreator(role)) {
+                if (isPremiumCreator(role, login)) {
                     usersAttemptsResponses.add(UsersAttemptsResponse.builder()
                         .userLogin(currentUserTest.getUserLogin())
                         .previousAttempts(previousAttempts.toArray(PreviousAttemptsResponse[]::new))
@@ -184,7 +197,7 @@ public class TestService {
                 userCount++;
             }
             sumMarks += userTest.getMark();
-            if (isPremiumCreator(role)) {
+            if (isPremiumCreator(role, login)) {
                 previousAttempts.add(PreviousAttemptsResponse.builder()
                     .dateTime(userTest.getDateTime())
                     .mark(userTest.getMark())
@@ -192,7 +205,7 @@ public class TestService {
             }
 
         }
-        if (currentUserTest != null && isPremiumCreator(role)) {
+        if (currentUserTest != null && isPremiumCreator(role, login)) {
             usersAttemptsResponses.add(UsersAttemptsResponse.builder()
                 .userLogin(currentUserTest.getUserLogin())
                 .previousAttempts(previousAttempts.toArray(PreviousAttemptsResponse[]::new))
@@ -207,14 +220,14 @@ public class TestService {
         testInfoResponse.setRole(role);
         testInfoResponse.setUserCount(userCount);
         testInfoResponse.setUserAverageMark(userAverageMark);
-        if (isPremiumCreator(role)) {
+        if (isPremiumCreator(role, login)) {
             testInfoResponse.setUsersAttemptsResponse(usersAttemptsResponses.toArray(UsersAttemptsResponse[]::new));
         }
         return testInfoResponse;
     }
 
-    private boolean isPremiumCreator(String role) {
-        return ARTICLE_CREATOR.equals(role) && isPremiumUser();
+    private boolean isPremiumCreator(String role, String login) {
+        return ARTICLE_CREATOR.equals(role) && isPremiumUser(login);
     }
 
     private PreviousAttemptsResponse[] getOwnPreviousAttempts(String login, int testId) {
@@ -284,13 +297,28 @@ public class TestService {
         userTestRepository.save(userTest);
     }
 
-    public void deleteTest(int id) {
+    public void deleteTest(String login, int id) {
+        checkAccessByTestId(login, id);
         testRepository.deleteById(id);
     }
 
-    //todo
-    private boolean isPremiumUser() {
-        return true;
+    private boolean isPremiumUser(String login) {
+        final Optional<User> optionalUser = userRepository.findById(login);
+        return optionalUser.map(user -> user.getRole()
+            .equals(PREMIUM_USER)).orElse(false);
+    }
 
+    private void checkAccessByArticleId(String login, Integer articleId) {
+        final UserArticle userArticle = userArticleRepository.findByUserLoginAndArticleId(login, articleId)
+            .orElseThrow(() -> new ServiceApiException(HttpStatus.FORBIDDEN));
+        if (!userArticle.getRole().equals(ARTICLE_CREATOR)) {
+            throw new ServiceApiException(HttpStatus.FORBIDDEN);
+        }
+    }
+
+    private void checkAccessByTestId(String login, Integer testId) {
+        final Test test = testRepository.findById(testId)
+            .orElseThrow(() -> new ServiceApiException(HttpStatus.NOT_FOUND));
+        checkAccessByArticleId(login, test.getArticleId());
     }
 }
